21567 숫자의 개수 2

+++ rust/src/21/21567.rs
@@ -2,7 +2,7 @@ use std::io;
 
 fn main() {
     let buf = io::read_to_string(io::stdin()).unwrap();
-    let product: i32 = buf.lines().map(|s| s.parse::<i32>().unwrap()).product();
+    let product: i64 = buf.lines().map(|s| s.parse::<i64>().unwrap()).product();
 
     for digit in '0'..='9' {
         println!("{}", product.to_string().matches(digit).count());
